Moved util functions to util file
